VME: Support acfail and sysfail interrupts
Both events are asynchronous and must usually be handled immediately,
so polling the corresponding status bits alone is not sufficient
(although it may be useful in certain cases). Extend the vme
framework to take care of these interrupts, and allow for registering
callback functions.

The irq_mtx mutex used for the generic IRQ callbacks is extended to
also cover manipulations of the interrupt masks in the acfail
and sysfail case.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>