vme_user: Make sysfs names more sensible
Specifying a path name in device_create does not
automatically generate the path, so change the
sysfs file names from bus/vme/mX (which is mapped
to bus!vme!mX) to b0mX, and similar for the slave
window and control files.

Currently, the driver is limited to one single bus,
but the name format allows for easily extending the
scheme to multiple buses.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>