IRQ generation: Ensure that we are not waiting for multiple IACKs
The tsi148 can not distinguish from the IACK which interrupt was
acknowledged. Consequently, if we generate a second interrupt while
the first one has not yet been acknowledged, we will consequently
not know which one was acknowledged once an IACK arrives.
We therefore need to ensure that at most one interrupt acknowledgement
is being waited for.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>