pwm: pca9685: Allow any of the 16 PWMs to be used as a GPIO
commit bccec89f0a35f65302734d1cdb01479df0f33ac9 upstream.

The PCA9685 controller has full on/off bit for each PWM channel. Setting
this bit bypasses the PWM control and the line works just as it would be a
GPIO. Furthermore in Intel Galileo it is actually used as GPIO output for
discreet muxes on the board.

This patch adds GPIO output only support for the driver so that we can
control the muxes on Galileo using standard GPIO interfaces available in
the kernel. GPIO and PWM functionality is exclusive so only one can be
active at a time on a single PWM channel.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Acked-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Thierry Reding <thierry.reding@gmail.com>
[Jan: adjusted gpio.parent to .dev for 4.4]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>