net: ethernet: ti: icss_iep: Update compare registers after changing ptp clock time
Consider the scenario where PPS is enabled, then iep set_time or adjust_time is
called with time deviation that is > PPS period. In this scenario there
is a possibility that PPS stopped as IEP compare registers are not
updated but the counter registers are updated and counter value can
never be matched with compare value.

To avoid this scenario, update the compare registers when set/adjust
time is called.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>